first stack changes (#14)
### TL;DR

Added a comment to the Exercise model.

### What changed?

Added a comment `#fist change to the models` at the end of the Exercise model class. This appears to be a documentation note (with a typo - "fist" instead of "first") marking this as the initial change to the model files.

### How to test?

No functional testing is required as this is only a comment addition. You can verify the comment appears in the Exercise model file.

### Why make this change?

This comment likely serves as a marker or reminder that this is the first modification made to the models in the application. It may be helpful for tracking the evolution of changes during development.