Second change (#15)
### TL;DR

Added a comment to the Routine model.

### What changed?

Added a comment `#second change to the models` at the end of the Routine class definition.

### How to test?

No functional testing is required as this is just a comment addition. You can verify the comment appears in the Routine model file.

### Why make this change?

This comment appears to be marking or documenting a second modification to the models. It may be used for tracking changes during development or as a reference point for future modifications.